Removed most of the warnings for Debug Unicode.
- 20 remaining warnings are in TinyXML code.
- A bunch of remaining warnings are CRT SECURITY warnings.

- Still need to clean up Debug non-Unicode and the release builds.

+++ PowerEditor/src/ScitillaComponent/DocTabView.cpp
@@ -71,7 +71,7 @@ BufferID DocTabView::findBufferByName(const TCHAR * fullfilename) {	//-1 if not 
 	TCITEM tie;
 	tie.lParam = -1;
 	tie.mask = TCIF_PARAM;
-	for(size_t i = 0; i < _nbItem; i++) {
+	for(int i = 0; i < _nbItem; i++) {
 		::SendMessage(_hSelf, TCM_GETITEM, i, reinterpret_cast<LPARAM>(&tie));
 		BufferID id = (BufferID)tie.lParam;
 		Buffer * buf = MainFileManager->getBufferByID(id);
@@ -86,7 +86,7 @@ int DocTabView::getIndexByBuffer(BufferID id) {
 	TCITEM tie;
 	tie.lParam = -1;
 	tie.mask = TCIF_PARAM;
-	for(int i = 0; i < (int)_nbItem; i++) {
+	for(int i = 0; i < _nbItem; i++) {
 		::SendMessage(_hSelf, TCM_GETITEM, i, reinterpret_cast<LPARAM>(&tie));
 		if ((BufferID)tie.lParam == id)
 			return i;
